Add product files and working scripts
Adds all the files for the products resource, as well as the curl
scripts to test them.

Products expects a name, category, price, description, image url, and
rating.

Products will need to be uploaded, modified and removed from an admin
account, as users will only be able to see them.