Add orders resource
Adds a resource to the back end for orders along with curl scripts.

Orders expects a date_placed, isOpen boolean, and a products array.
When a new order is created, the date_placed will be null, isOpen will
be true, and products will be empty.

As products are added, the products array will fill up. When an
order is checked out, isOpen will be set to false and date_placed
will be set to the current date at the time the order was placed.