Stripe, that pig, is working
Ok so adding code that nick and I have been working on together, we
finally got our request to work.

Our problem was that we didn't have api access to Stripe, so we
added a direct secret key to each stripe request. We originally
thought that it included it in the stripe require but that is a pig
hearted lie.

Boom baby.