Created a new param for orders
This is a false param that wlll only be populated upon order completion
with the proof of purchase installed into the orders table. This only
shows the id stripe consumer id, the amount paid, and the status
that was achieved on it. No sensitive user data is located on it at all.