Updated backend
So the backend now has a lot going for it. The product model now has an
attribute called status that, when and a product is created, makes is
set to a default. This status then changes to hidden when the user
decides to update on the front end. The hidden aspect is called upon
within handlebars code to enable the product to be hidden in the store
and for allowing of filtering on loading cart page.

There are still attempted curl requests and admin toggle functions that
I am leaving in for showing and confirming that I attempted to let a
user be changed by another admin. I have come to the conclusion that
that is inefficient and needless as I should have only one person be
in charge of that instead of any admin.